Problem
This is a linux solution for #62
When building with ccache, the ccache executable is seen to crash with
the error

ccache: invalid option -- 't'

Solution

The issue as described at
https://github.com/mapnik/python-mapnik/issues/88 is that the
environmental variable CXX on linux will only accpet the first space
separated sting in that variable as the linker when compiling sources in
python.

A work around is to create a shell script that is used as a shim for the
linker, it calls ccache <linker> and as the script name is one command
the linking will succeed.

Result

Compilation of the C++ code will link on linux