INSTALLATION: Tox whitelist should match the virt env in the docs (#135)
* INSTALLATION: Tox whitelist should match the virt env in the docs

Problem

The tox linter has a whilelist of directories which will not get
scanned. The virtual environment that is whitelisted does not match the
virtual env created in the INSTALLATION doc

Solution

Change the virtual environement creation instructions to match the
directory in the tox configuration.

* INSTALLATION: Tox whitelist should match the virt env in the docs

Problem

The tox linter has a whilelist of directories which will not get
scanned. The virtual environment that is whitelisted does not match the
virtual env created in the INSTALLATION doc

Solution

Change the virtual environement creation instructions to match the
directory in the tox configuration.

* Adding vendors directory to list of excluded directories to lint as we
have no direct control on t3rd party repos

* Remove ignore baseline conflict flag