Allowing user to load plugin with given sample rate and max block size
Problem

This fixes issue #399. Some plugins require the sample rate to the
established already at init, where they perform sanity checks.
To allow for more flexibility and for the sake of consistency, we
can do the same for the block size.

Solution

Add sample_rate and maximum_block_size as optional input arguments
to load_plugin(). The default values are kept to the ones that were
previously hardcoded.